More complete way to get a version number
We now also add -dirty when the working directory contains changes to the last
commit, and cache the result until a source file is modified again or a git
commit happened. A new git tag will also trigger the version to be updated.

The version file needn't being copied explicitely now , as it ends up in the
MAKEFILE_LIST variable.

Note that the detection of a new commit relies on git reflog being enabled,
which is the default. Nobody in their full mind should disable reflog anyways.