GNUmakefile: Allow sources file under subdirectories to be built
For now, only building programs, with a forced dependency on libvmfs.a is
supported. Any .c file in the subdirectory is considered a source file.
The program name is the directory name. The CFLAGS for each source file
can be tuned with the source.c_CFLAGS variable defined in manifest.mk.

Only sub directories with a manifest.mk file are considered.