Normalize entry and item for fbb
In fbb case, entry as returned by vmfs_block_get_info() is always 0, despite
the defined number of items per entry not being infinite. In such case, and
only for the function displaying the entry and item corresponding to a given
blkid in debugvmfs, we normalize entry and item so that they fit the defined
number of items per entry in the file block bitmap.