Add support for "inline" content
In VMFS 5, content of files which size is smaller than 1KB are stored inline
in the inode block list. Such inodes apparently have a ZLA fixed to 0x10d1.
With the block_id construct, this is type 1 (file block), flag 2 (unused so far),
entry 0, item 67 (0x43). We assume only the flag is significant.